fix: solve logic

+++ src/test/scala/kiln/math/PermutationSuite.scala
@@ -7,10 +7,10 @@ class PermutationSuite extends munit.FunSuite:
 
   test("given n = 0, count returns 1"):
     val s = Permutation(0)
-    val expected: BigInt = 1
+    val expected: Int = 1
     assertEquals(s.count, expected)
 
   test("given n = 3, count answers the amount of permutations (6)"):
     val s = Permutation(3)
-    val expected: BigInt = 6
+    val expected: Int = 6
     assertEquals(s.count, expected)
